+++ plugin-build/src/test/kotlin/io/sentry/android/gradle/SentryPluginAutoInstallTest.kt
@@ -43,6 +43,8 @@ class SentryPluginAutoInstallTest(
               implementation 'androidx.fragment:fragment:1.3.5'
               // our plugin shouldn't install okhttp, since it's a direct dep
               implementation 'io.sentry:sentry-android-okhttp:5.4.0'
+              // our plugin shouldn't install sqlite, since it's a direct dep
+              implementation 'io.sentry:sentry-android-sqlite:6.21.0'
             }
 
             sentry {
@@ -58,6 +60,8 @@ class SentryPluginAutoInstallTest(
         assertTrue { "io.sentry:sentry-android-fragment:5.1.0" in result.output }
         assertFalse { "io.sentry:sentry-android-okhttp:5.1.0" in result.output }
         assertTrue { "io.sentry:sentry-android-okhttp:5.4.0" in result.output }
+        assertFalse { "io.sentry:sentry-android-sqlite:5.1.0" in result.output }
+        assertTrue { "io.sentry:sentry-android-sqlite:6.21.0" in result.output }
         assertFalse { "io.sentry:sentry-compose-android:5.1.0" in result.output }
 
         // ensure all dependencies could be resolved
@@ -85,6 +89,7 @@ class SentryPluginAutoInstallTest(
         assertFalse { "io.sentry:sentry-android-timber:$SENTRY_SDK_VERSION" in result.output }
         assertFalse { "io.sentry:sentry-android-fragment:$SENTRY_SDK_VERSION" in result.output }
         assertFalse { "io.sentry:sentry-android-okhttp:$SENTRY_SDK_VERSION" in result.output }
+        assertFalse { "io.sentry:sentry-android-sqlite:$SENTRY_SDK_VERSION" in result.output }
 
         // ensure all dependencies could be resolved
         assertFalse { "FAILED" in result.output }
@@ -158,6 +163,49 @@ class SentryPluginAutoInstallTest(
         val result = runListDependenciesTask()
 
         assertTrue { "io.sentry:sentry-compose-android:6.7.0" in result.output }
+        // ensure all dependencies could be resolved
+        assertFalse { "FAILED" in result.output }
+    }
+
+    @Test
+    fun `sqlite is not added for lower sentry versions`() {
+        appBuildFile.appendText(
+            // language=Groovy
+            """
+            dependencies {
+              implementation 'androidx.sqlite:sqlite:2.0.0'
+            }
+
+            sentry.autoInstallation.enabled = true
+            sentry.autoInstallation.sentryVersion = "6.20.0"
+            sentry.includeProguardMapping = false
+            """.trimIndent()
+        )
+
+        val result = runListDependenciesTask()
+
+        assertFalse { "io.sentry:sentry-android-sqlite:6.20.0" in result.output }
+        assertFalse { "FAILED" in result.output }
+    }
+
+    @Test
+    fun `sqlite is added with when sentry version 6_21_0 or above is used`() {
+        appBuildFile.appendText(
+            // language=Groovy
+            """
+            dependencies {
+              implementation 'androidx.sqlite:sqlite:2.0.0'
+            }
+
+            sentry.autoInstallation.enabled = true
+            sentry.autoInstallation.sentryVersion = "6.21.0"
+            sentry.includeProguardMapping = false
+            """.trimIndent()
+        )
+
+        val result = runListDependenciesTask()
+
+        assertTrue { "io.sentry:sentry-android-sqlite:6.21.0" in result.output }
         // ensure all dependencies could be resolved
         assertFalse { "FAILED" in result.output }
     }